Clean up, reworked fetches
